requirements.txt
asgiref==3.9.1
coloredlogs==15.0.1
Django==4.2.23
flatbuffers==25.2.10
humanfriendly==10.0
image==1.5.33
imageio==2.37.0
lazy_loader==0.4
mpmath==1.3.0
networkx==3.2.1
nudenet==3.4.2
NudeNetClassifier @ git+https://github.com/platelminto/NudeNet@05ad8a29b4a2474e6cc4f7f76c76e77e01b08744
numpy==2.0.2
onnxruntime==1.19.2
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
protobuf==6.31.1
scikit-image==0.24.0
scipy==1.13.1
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
tifffile==2024.8.30
typing_extensions==4.14.1
